refactor(config): split display/redaction helpers into display.rs
Keep crates/config/src/lib.rs under the 500-line file limit (552 -> 440) by
moving REDACTED_CONFIG_KEYS, redact_if_sensitive, and build_config_entries into
a new display module; build_config_entries is re-exported so the public path is
unchanged.
